feat: Update footer with social media links and contact info
Linked social media icons in the footer to respective profiles.

Updated the third column in the footer to 'Call Us' and added a phone number.

Changed the second column in the footer to 'Group Companies' and added random company names.

TODO: Link list elements in the 'Company' column to their respective pages.

TODO: Implement routing in the navbar to different pages.

TODO: Add dynamic searching based on product name in the navbar.